--- tutorial/src/main/scala/littleMonadDemo/ch07Function1Monad/Example07.scala
@@ -1,30 +1,0 @@
-package littleMonadDemo.ch07Function1Monad
-
-import scala.util.chaining._
-
-import littleMonadDemo.libMyCats._
-import littleMonadDemo.libCompute.WithMyCats._
-
-// uncomment for use with Cats instead of libMyCats
-// import cats._
-// import cats.data._
-// import cats.implicits._
-// import littleMonadDemo.libCompute.WithCats._
-
-object Example07 extends util.App {
-
-  val plus1: Int => Int  = (x => x + 1)
-  val times2: Int => Int = (x => x * 2)
-
-  val fn1: Function1[Int, (Int, Int)] = for {
-    i1 <- plus1
-    i2 <- times2
-  } yield (i1, i2)
-
-  fn1(10) pipe println
-
-  compute(plus1, times2).apply(10) pipe println // (11, 20)
-
-  val fn2: Int => (Int, Int) = compute(plus1, times2)
-  fn2(10) pipe println // (11, 20)
-}
