+++ tutorial/src/main/scala/tutorial/ch03IdentityMonad/Example03.scala
@@ -1,0 +1,11 @@
+package tutorial.ch03IdentityMonad
+
+import tutorial.libCompute.WithMyCats._
+
+object Example03 extends util.App {
+
+  println("----- Identity:")
+
+  val idResult = compute(42, 42)
+  println(idResult)
+}
