+++ tutorial/src/main/scala/tutorial/ch04EitherMonad/README.md
