+++ tutorial/src/main/scala/tutorial/ch06FunctionsAreValues/README.md
