+++ tutorial/src/main/scala/tutorial/ch09IOMonad/IOAppHelloWithEval.scala
@@ -1,0 +1,19 @@
+package tutorial.ch09IOMonad
+
+import tutorial.libMyCats._
+
+object IOAppHelloWithEval extends util.App {
+
+  // referentially transparent description of the program
+
+  val hello: IO[Unit] = for {
+    _    <- IO.eval(print("What's your name?  "))
+    name <- IO.eval(scala.io.StdIn.readLine())
+    _    <- IO.eval(println(s"Hello $name!\n"))
+  } yield ()
+
+  // interpretation / execution of the program
+  // is NOT referentially transparent
+
+  hello.unsafeRun()
+}
